refactor: move test helpers into separate file
